Add test case for coords missing zero-padding
Uses a JWKS from our `sandbox` environment, which we know is missing
zero-padding for one of the EC key coords. Also uses a webhook payload
and signature, from a manual test done in `sandbox`.

This test case proves that the `jwt` gem isn't able to successfully
import the EC key, while the `truelayer-signing` library handles it just
fine and is able to verify the signature.