Python: publish on new tag only
Since #228, the Python workflow will run for any changes in the Python
code, the shared test resources _and_ the Python workflow itself. This
means that when the library version does not change, the `publish` step
simply fails, which does not look good.

This updates the Python workflow so that it will only attempt to publish
the library when there is a `python/v**` tag being pushed. It does mean
that it now requires an additional manual step, but it also means that
we can more easily batch changes in each version bump, without the
workflow failing in between.