Python: publish on version bump only (#230)
Since #228, the Python workflow will run for any changes in the Python
code, the shared test resources _and_ the Python workflow itself. This
means that when the library version does not change, the `publish` step
simply fails, which does not look good.

This updates the Python workflow so that it will only attempt to publish
the library automatically when there is a version bump on `main`.